requirements.txt
requests
nbconvert
nbformat
python-dotenv
pyyaml
pytest